onchain-analytics/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.3
bitstring==4.2.3
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
crc16==0.1.1
Deprecated==1.2.14
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.2
idna==3.10
iniconfig==2.0.0
loguru==0.7.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
propcache==0.2.0
pybit==5.8.0
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyfiglet==1.0.2
PyNaCl==1.5.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytonapi==0.4.2
pytoncenter==0.0.14
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
tomli==2.1.0
ton==0.26
tonpy==0.0.0.1.2b0
tonsdk==1.0.15
treelib==1.7.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.27.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
